requirements.txt
# GIO Crypto Bot - Зависимости
# Основные библиотеки
asyncio-mqtt==0.16.1
aiohttp==3.9.1
aiosqlite==0.19.0
APScheduler==3.10.4
numpy==1.24.3
pandas==2.0.3
websockets==11.0.3
psutil==5.9.6

# Криптографические библиотеки
cryptography==41.0.7
pycryptodome==3.19.0

# Работа с данными
python-dotenv==1.0.0
ujson==5.8.0
orjson==3.9.10

# Логирование и мониторинг
coloredlogs==15.0.1
structlog==23.2.0

# Валидация и парсинг
pydantic==2.5.0
python-dateutil==2.8.2

# HTTP клиенты
httpx==0.25.2
requests==2.31.0

# Анализ данных
scipy==1.11.4
scikit-learn==1.3.2
ta-lib==0.4.28  # Опционально для дополнительных индикаторов

# Утилиты
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Тестирование (опционально)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Разработка (опционально)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# ������������������ � �����������
psutil>=5.9.0
numba>=0.58.0
memory-profiler>=0.60.0

telegram-bot==20.7
